first-resume-parser/requirements.txt
pdfplumber
python-docx
pandas
google-genai
python-dotenv
beautifulsoup4
streamlit